otelcol.receiver.file_stats: shim on unsupported platforms
A bug in the upstream receiver prevents the component from compiling
properly on FreeBSD. As a temporary workaround, we shim out
implementations on unsupported platforms and return a fake component
which is always unhealthy.
