+++ internal/component/otelcol/receiver/file_stats/types_convert.go
@@ -1,0 +1,110 @@
+//go:build linux || darwin || windows
+
+package file_stats
+
+import (
+	"github.com/grafana/alloy/internal/component/otelcol"
+	"github.com/grafana/alloy/internal/component/otelcol/receiver"
+	"github.com/mitchellh/mapstructure"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/filestatsreceiver"
+	otelcomponent "go.opentelemetry.io/collector/component"
+	otelextension "go.opentelemetry.io/collector/extension"
+)
+
+var _ receiver.Arguments = Arguments{}
+
+// Convert implements receiver.Arguments.
+func (args Arguments) Convert() (otelcomponent.Config, error) {
+	var out filestatsreceiver.Config
+
+	out.ControllerConfig = *args.Controller.Convert()
+	out.Include = args.Include
+
+	// We have to use mapstructure.Decode for args.MetricsBuilder because the
+	// upstream type is in an internal package.
+	err := mapstructure.Decode(
+		args.MetricsBuilder.toMap(),
+		&out.MetricsBuilderConfig,
+	)
+
+	return &out, err
+}
+
+// Extensions implements receiver.Arguments.
+func (args Arguments) Extensions() map[otelcomponent.ID]otelextension.Extension {
+	return nil
+}
+
+// Exporters implements receiver.Arguments.
+func (args Arguments) Exporters() map[otelcomponent.DataType]map[otelcomponent.ID]otelcomponent.Component {
+	return nil
+}
+
+// NextConsumers implements receiver.Arguments.
+func (args Arguments) NextConsumers() *otelcol.ConsumerArguments {
+	return args.Output
+}
+
+// DebugMetricsConfig implements receiver.Arguments.
+func (args Arguments) DebugMetricsConfig() otelcol.DebugMetricsArguments {
+	return args.DebugMetrics
+}
+
+// toMap encodes args to a map for use with mapstructure.Decode.
+func (args *MetricsBuilderArguments) toMap() map[string]any {
+	return map[string]any{
+		"metrics":             args.Metrics.toMap(),
+		"resource_attributes": args.ResourceAttributes.toMap(),
+	}
+}
+
+// toMap encodes args to a map for use with mapstructure.Decode.
+func (args *MetricsArguments) toMap() map[string]any {
+	return map[string]any{
+		"file.atime": args.FileAtime.toMap(),
+		"file.count": args.FileCount.toMap(),
+		"file.ctime": args.FileCtime.toMap(),
+		"file.mtime": args.FileMtime.toMap(),
+		"file.size":  args.FileSize.toMap(),
+	}
+}
+
+// toMap encodes args to a map for use with mapstructure.Decode.
+func (args *MetricArguments) toMap() map[string]any {
+	return map[string]any{"enabled": args.Enabled}
+}
+
+// toMap encodes args to a map for use with mapstructure.Decode.
+func (args *ResourceAttributesArguments) toMap() map[string]any {
+	return map[string]any{
+		"file.name": args.FileName.toMap(),
+		"file.path": args.FilePath.toMap(),
+	}
+}
+
+// toMap encodes args to a map for use with mapstructure.Decode.
+func (args *ResourceAttributeArguments) toMap() map[string]any {
+	includes := make([]map[string]any, 0, len(args.MetricsInclude))
+	excludes := make([]map[string]any, 0, len(args.MetricsExclude))
+
+	for _, include := range args.MetricsInclude {
+		includes = append(includes, include.toMap())
+	}
+	for _, exclude := range args.MetricsExclude {
+		excludes = append(excludes, exclude.toMap())
+	}
+
+	return map[string]any{
+		"enabled":         args.Enabled,
+		"metrics_include": includes,
+		"metrics_exclude": excludes,
+	}
+}
+
+// toMap encodes args to a map for use with mapstructure.Decode.
+func (args *FilterArguments) toMap() map[string]any {
+	return map[string]any{
+		"strict": args.Strict,
+		"regexp": args.Regex,
+	}
+}
